+++ kuma-client/src/config.rs
@@ -51,6 +51,9 @@ pub struct Config {
     /// The MFA token for logging into Uptime Kuma (required if MFA is enabled).
     pub mfa_token: Option<String>,
 
+    /// The MFA secret. Used to generate a tokens for logging into Uptime Kuma (alternative to a single_use mfa_token).
+    pub mfa_secret: Option<String>,
+
     /// List of HTTP headers to send when connecting to Uptime Kuma.
     #[serde_as(
         as = "PickFirst<(DeserializeVecLenient<String>, StringWithSeparator::<CommaSeparator, String>)>"
@@ -77,6 +80,7 @@ impl Default for Config {
             username: None,
             password: None,
             mfa_token: None,
+            mfa_secret: None,
             headers: Vec::new(),
             connect_timeout: 30.0,
             call_timeout: 30.0,