+++ kuma-client/src/error.rs
@@ -1,3 +1,5 @@
+use std::time::SystemTimeError;
+
 use thiserror::Error;
 
 /// Custom error type for handling various errors in the kuma_client library.
@@ -26,6 +28,10 @@ pub enum Error {
     /// Error when the server expects a username/password, but none was provided.
     #[error("It looks like the server is expecting a username/password, but none was provided")]
     NotAuthenticated,
+
+    /// Error when mfa is enabled, but no token/secret was provided.
+    #[error("MFA enabled, but no token/secret was provided")]
+    TokenRequired,
 
     /// Connection loss to Uptime Kuma.
     #[error("Connection to Uptime Kuma was lost")]
@@ -66,7 +72,34 @@ pub enum Error {
     /// Wrapper for an underlying reqwest error.
     #[error(transparent)]
     Reqwest(#[from] reqwest::Error),
+
+    /// Wrapper for an underlying totp_rs error.
+    #[error(transparent)]
+    Totp(#[from] TotpError),
 }
 
 /// Custom result type for handling various errors in the kuma_client library.
 pub type Result<T> = std::result::Result<T, Error>;
+
+/// Wrapper error type for totp_rs errors.
+#[derive(Error, Debug)]
+pub enum TotpError {
+    /// Wrapper for an underlying totp_rs::TotpUrlError error.
+    #[error(transparent)]
+    TotpUrlError(#[from] totp_rs::TotpUrlError),
+
+    /// Wrapper for an underlying totp_rs::SecretParseError error.
+    #[error(transparent)]
+    SecretParseError(#[from] totp_rs::SecretParseError),
+
+    /// Wrapper for an underlying totp_rs::Rfc6238Error error.
+    #[error(transparent)]
+    Rfc6238Error(#[from] totp_rs::Rfc6238Error),
+
+    /// Wrapper for an underlying SystemTimeError error.
+    #[error(transparent)]
+    SystemTimeError(#[from] SystemTimeError),
+}
+
+/// Custom result type for handling totp_rs errors.
+pub type TotpResult<T> = std::result::Result<T, TotpError>;