+++ kuma-client/src/models/response.rs
@@ -3,11 +3,18 @@ use serde_with::skip_serializing_none;
 
 #[skip_serializing_none]
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct LoginResponse {
-    #[serde(rename = "ok")]
-    pub ok: bool,
-    #[serde(rename = "msg")]
-    pub msg: Option<String>,
-    #[serde(rename = "token")]
-    pub token: Option<String>,
+#[serde(untagged)]
+pub enum LoginResponse {
+    Normal {
+        #[serde(rename = "ok")]
+        ok: bool,
+        #[serde(rename = "msg")]
+        msg: Option<String>,
+        #[serde(rename = "token")]
+        token: Option<String>,
+    },
+    TokenRequired {
+        #[serde(rename = "tokenRequired")]
+        token_required: bool,
+    },
 }